+++ src/main/java/com/aaax/server/config/security/AuthenticationServerConfig.java
@@ -450,7 +450,7 @@ public class AuthenticationServerConfig {
         RSAPrivateKey privateKey = (RSAPrivateKey) keyPair.getPrivate();
         return new RSAKey.Builder(publicKey)
                 .privateKey(privateKey)
-                .keyID("altech-uaa")
+                .keyID("aaax")
                 .keyUse(KeyUse.SIGNATURE)
                 .build();
     }