+++ src/main/java/com/aaax/server/listener/UserDeletedNotificationListener.java
@@ -2,12 +2,11 @@ package com.aaax.server.listener;
 
 
 import com.aaax.core.api.DiscordApiClient;
-import com.aaax.core.api.dto.DiscordWebhookMessage;
+import com.aaax.core.api.DiscordWebhookSupport;
 import com.aaax.core.kafka.BaseListener;
 import com.aaax.core.kafka.event.UserDeletedEvent;
 import com.aaax.core.utils.JSONUtil;
 import com.aaax.core.utils.KafkaUtil;
-import com.aaax.core.utils.RetrofitCallHandler;
 import com.aaax.server.usecase.UserManagementUseCase;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -31,9 +30,9 @@ public class UserDeletedNotificationListener extends BaseListener {
     private final UserManagementUseCase userManagementUseCase;
     @Value("${config.system-invoker}")
     protected String systemInvoker;
-    @Value("${ext.api.client.discord.webhookId}")
+    @Value("${ext.api.client.discord.webhookId:}")
     private String webhookId;
-    @Value("${ext.api.client.discord.webhookToken}")
+    @Value("${ext.api.client.discord.webhookToken:}")
     private String webhookToken;
 
     // ______ this is the execute method you need to better implemented
@@ -52,10 +51,7 @@ public class UserDeletedNotificationListener extends BaseListener {
     }
 
     private void discord(ConsumerRecord<String, String> payload) {
-        DiscordWebhookMessage message = DiscordWebhookMessage.builder()
-                .content("```".concat(payload.value()).concat("```"))
-                .build();
-        RetrofitCallHandler._void_execute(discordApiClient.sendWebhookMessage(message, webhookId, webhookToken));
+        DiscordWebhookSupport.sendSafe(discordApiClient, webhookId, webhookToken, "```".concat(payload.value()).concat("```"));
     }
 
     @KafkaListener(